+++ app/.eslintignore
@@ -1,0 +1,1 @@
+create_component
