+++ app/.storybook/main.ts
@@ -2,12 +2,7 @@ import type { StorybookConfig } from '@storybook/nextjs'
 import path from 'path'
 
 const config: StorybookConfig = {
-    stories: [
-        {
-            directory: '../src/',
-            files: '**/*.@(mdx|stories.@(js|jsx|mjs|ts|tsx))',
-        },
-    ],
+    stories: ['../src'],
     addons: [
         '@storybook/addon-links',
         '@storybook/addon-essentials',
