+++ app/create_component
@@ -1,8 +1,7 @@
 #!/bin/zsh
 for name in $@; do
     cp -R ./src/components/template ./src/components/$name
-    Name=$(tr '[:lower:]' '[:upper:]' <<<${name:0:1})${name:1}
-    cat ./src/components/$name/Template.tsx | sed -e "s/Template/$Name/g" >./src/components/$name/$Name.tsx
-    cat ./src/components/$name/Template.stories.tsx | sed -e "s/Template/$Name/g" >./src/components/$name/$Name.stories.tsx
+    cat ./src/components/$name/Template.tsx | sed -e "s/Template/$name/g" >./src/components/$name/$name.tsx
+    cat ./src/components/$name/Template.stories.tsx | sed -e "s/Template/$name/g" >./src/components/$name/$name.stories.tsx
     rm ./src/components/$name/Template*.tsx
 done
