+++ app/src/app/blog/[slug]/page.tsx
@@ -1,8 +1,8 @@
 import fs from 'fs'
 
 import { Blocks } from '@/components/block/Block'
-import Profile from '@/components/common/profile/Profile'
-import PageInfo from '@/components/page/blog/pageInfo/PageInfo'
+import Profile from '@/components/common/Profile/Profile'
+import PageInfo from '@/components/page/blog/PageInfo/PageInfo'
 import { Page as PageData } from '@/type/page/page'
 
 const NEXT_PUBLIC_URL = process.env.NEXT_PUBLIC_URL
