+++ app/src/app/layout.tsx
@@ -1,6 +1,8 @@
-import Header from '@/components/layout/Header'
+import Header from '@/components/layout/Header/Header'
 
 import type { Metadata } from 'next'
+
+
 import './globals.css'
 
 export const metadata: Metadata = {
