--- app/src/components/block/Header.stories.tsx
@@ -1,32 +1,0 @@
-import Header from '../layout/Header'
-
-import type { Meta, StoryObj } from '@storybook/react'
-
-// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
-const meta = {
-    component: Header,
-    parameters: {
-        // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
-        layout: 'centered',
-    },
-    // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
-    tags: ['autodocs'],
-    // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
-    argTypes: {},
-} satisfies Meta<typeof Header>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-export const Default: Story = {
-    args: {},
-    render() {
-        return (
-            <div>
-                <meta.component></meta.component>
-                <main>{new Array(1000).fill(0).map((v, i) => `${i} `)}</main>
-            </div>
-        )
-    },
-}
