+++ app/src/components/block/bookMark/BookMark.stories.tsx
@@ -1,10 +1,10 @@
-import Equation from '../equation/Equation'
+import BookMark from './BookMark'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    component: Equation,
+    component: BookMark,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -13,13 +13,11 @@ const meta = {
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
     argTypes: {},
-} satisfies Meta<typeof Equation>
+} satisfies Meta<typeof BookMark>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
-    args: {
-        math: 'x^2 + y^2 = z^2',
-    },
+    args: { href: 'https://www.google.com' },
 }
