+++ app/src/components/block/code/Code.tsx
@@ -1,11 +1,11 @@
-"use client"
+'use client'
+import './Code.style.scss'
 
 import { CSSProperties, useEffect, useRef, useState } from 'react'
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter'
 import { vscDarkPlus } from 'react-syntax-highlighter/dist/cjs/styles/prism'
 
-import Tooltip from '../../common/tooltip/Tooltip'
-import './Code.style.scss'
+import Tooltip from '../../common/Tooltip/Tooltip'
 
 interface CodeProps {
     children: string
