+++ app/src/components/layout/Header/Header.stories.tsx
@@ -1,10 +1,10 @@
-import { Button } from './button/Button'
+import Header from './Header'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-    component: Button,
+    component: Header,
     parameters: {
         // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
         layout: 'centered',
@@ -12,33 +12,21 @@ const meta = {
     // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
     tags: ['autodocs'],
     // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
-    argTypes: {
-        mode: {
-            control: 'select',
-            options: ['primary', 'secondary', 'orange'],
-        },
-    },
-} satisfies Meta<typeof Button>
+    argTypes: {},
+} satisfies Meta<typeof Header>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 // More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-export const Primary: Story = {
-    args: {
-        mode: 'primary',
-        children: 'Button',
-    },
-}
-export const Secondary: Story = {
-    args: {
-        mode: 'secondary',
-        children: 'Button',
-    },
-}
-export const Orange: Story = {
-    args: {
-        mode: 'orange',
-        children: 'Button',
+export const Default: Story = {
+    args: {},
+    render() {
+        return (
+            <div>
+                <meta.component></meta.component>
+                <main>{new Array(1000).fill(0).map((v, i) => `${i} `)}</main>
+            </div>
+        )
     },
 }
